pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "security-threat-analyzer"
version = "1.0.0"
description = "AI-powered security log analysis using RAG (Retrieval-Augmented Generation)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "Vallabh SG"},
]

dependencies = [
    "langchain>=1.2.0",
    "langchain-core>=1.2.0",
    "langchain-community>=0.4.0",
    "langchain-ollama>=1.0.0",
    "langchain-text-splitters>=1.1.0",
    "langchain-huggingface>=1.2.0",
    "langchain-classic>=0.1.0",
    "faiss-cpu>=1.13.0",
    "sentence-transformers>=2.6.0",
    "streamlit>=1.39.0",
    "ollama>=0.6.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "requests>=2.32.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4.0",
    "mypy>=1.10",
]

[project.scripts]
security-analyzer = "security_analyzer.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
